app/requirements.txt
dagster
dagit
dagster-docker
dagster-k8s
dagster-aws
pandas
wordcloud
matplotlib
psycopg2-binary
sqlalchemy